% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_cov_coef_corel_dens_com.R
\name{cov_coef_corel}
\alias{cov_coef_corel}
\title{Calculul covariantei si coeficientului de corelatie pentru doua variabile aleatoare continue}
\usage{
cov_coef_corel(f_xy, lo_x, hi_x, lo_y, hi_y)
}
\arguments{
\item{f_xy}{Densitatea comuna de probabilitate a doua variabile aleatoare continue.}

\item{lo_x}{Valoarea inferioara a domeniului functiei fx}

\item{hi_x}{Valoarea superioara a domeniului functiei fx}

\item{lo_y}{Valoarea inferioara a domeniului functiei fy}

\item{hi_y}{Valoarea superioara a domeniului functiei fy}
}
\value{
Covarianta \code{cov}, coeficientul de corelatie \code{corr}.
}
\description{
Calculul covariantei si coeficientului de corelatie pentru doua variabile aleatoare continue
}
\examples{
cov_coef_corel(f, 0, 1, 0, 1)
}
