% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_dens_marg_cond.R
\name{dens_marg_cond}
\alias{dens_marg_cond}
\title{Pornind de la densitatea comuna a doua variabile aleatoare continue,
construirea densitatilor marginale si a densitatilor conditionate}
\usage{
dens_marg_cond(f_xy, lo_x, hi_x, lo_y, hi_y)
}
\arguments{
\item{f}{Densitatea comuna de probabilitate a doua variabile aleatoare continue.}
}
\value{
Se afiseaza cele doua densitati marginale si cele doua densitati conditionate.
}
\description{
Pornind de la densitatea comuna a doua variabile aleatoare continue,
construirea densitatilor marginale si a densitatilor conditionate
}
\examples{
dens_marg_cond(f, 0, 1, 0, 1)
}
