% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_med_dis_funct_cont.R
\name{medie_dispersie_func_cont}
\alias{medie_dispersie_func_cont}
\title{Calculul mediei si dispersiei unei variabile aleatoare g(X),
unde X are o repartisie continua cunoscuta, iar g este o functie
continua precizata de utilizator}
\usage{
medie_dispersie_func_cont(f)
}
\arguments{
\item{f}{Densitatea de probabilitate a variabilei aleatoare continue.}
}
\value{
Media \code{medie}, dispersia \code{dispersie}
}
\description{
Calculul mediei si dispersiei unei variabile aleatoare g(X),
unde X are o repartisie continua cunoscuta, iar g este o functie
continua precizata de utilizator
}
\examples{
medie_dispersie_func_cont(f)
}
