% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_med_dis_mem_init_cent_ord4.R
\name{medie_dispersie_momente}
\alias{medie_dispersie_momente}
\title{Acesta functie efectueaza calculul mediei, dispersiei si a momentelor
initiale si centrate pana la ordinul 4(daca exista). Atunci cand unul dintre
momente nu exista, se va afisa un mesaj corespunzator catre utilizator}
\usage{
medie_dispersie_momente(f)
}
\arguments{
\item{f}{Densitatea de probabilitate a variabilei aleatoare continue.}
}
\value{
Media \code{medie}, dispersia \code{dispersie},
lista de momente initiale \code{vector_mom_initial},
lista de momente centrate \code{vector_mom_centrate}.
}
\description{
Acesta functie efectueaza calculul mediei, dispersiei si a momentelor
initiale si centrate pana la ordinul 4(daca exista). Atunci cand unul dintre
momente nu exista, se va afisa un mesaj corespunzator catre utilizator
}
\examples{
medie_dipsersie_momente(f)
}
