\name{sum_diff}
\alias{sum_diff}

\title{
SUM_DIFF
}
\description{
The function calculates the sum or difference between two independent continuous random variaables.
}
\usage{
convolution_sum(fx, fy)
convolution_diff(fx, fy)
}

\details{
Calculation is done using the convolution theory.
}
\value{
Plots two graphs.
}

\author{
Bb4k
}
