% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_dens_marg_cond.R
\name{dens_marg_cond}
\alias{dens_marg_cond}
\title{Calculul covariantei si a coeficientului de
corelatie pentru doua variabile aleatoare continue}
\usage{
dens_marg_cond(f_xy, lo_x, hi_x, lo_y, hi_y)
}
\arguments{
\item{f}{Densitatea de probabilitate a variabilei aleatoare continue.}
}
\value{
Media \code{medie}, dispersia \code{dispersie},
lista de momente initiale \code{vector_mom_initial},
lista de momente centrate \code{vector_mom_centrate}.
}
\description{
Calculul covariantei si a coeficientului de
corelatie pentru doua variabile aleatoare continue
}
\examples{
medie_dipsersie_momente(f)
}
